+++ RunningRacesApi/Enums/WayPointMatchStatus.cs
@@ -1,0 +1,7 @@
+﻿namespace RunningRacesApi.Enums;
+public enum WayPointMatchStatus
+{
+    Exact,
+    Partial,
+    NotFound
+}