+++ RunningRacesApi/Services/IWayPointService.cs
@@ -5,6 +5,7 @@ namespace RunningRacesApi.Services;
 public interface IWayPointService
 {
     Task<PagedResult<WayPoint>> GetAllAsync(BaseSearchModel? searchModel = null);
+    Task<WayPoint> GetByIdAsync(int? id);
     Task<WayPoint> CreateAsync(WayPoint wayPoint);
     Task<WayPoint?> UpdateAsync(int id, WayPoint wayPoint);
     Task<bool> DeleteAsync(int id);